Moved math regression test to be true unit tests.
Moved matrix operations out of the quaternion operations file for better organization
