+++ tests/unit_tests/regression/CMakeLists.txt
@@ -6,7 +6,6 @@ target_sources(
     test_solver.cpp
     test_rotating_beam.cpp
     test_cantilever_beam.cpp
-    test_math.cpp
     test_rotor.cpp
     test_utilities.cpp
     test_controller.cpp
